SP int: make used and size fields unsigned
used and size have no reason to be negative - change type.
Change code to match unsigned change. Mostly change variables to be
unsigned where possible.
integer.c: Only have mp_rand_prime available when needed and
mp_prime_is_prime_ex is available.
Fixes from regression testing.

+++ wolfcrypt/src/asn.c
@@ -35590,7 +35590,7 @@ static int ParseCRL_Extensions(DecodedCRL* dcrl, const byte* buf,
 
                         if (ret == 0) {
                             dcrl->crlNumber = 0;
-                            for (i = 0; i < (*m).used; ++i) {
+                            for (i = 0; i < (int)(*m).used; ++i) {
                                 if (i > (CHAR_BIT *
                                          (int)sizeof(word32) / DIGIT_BIT)) {
                                     break;
